+++ install-look.sh
@@ -3,7 +3,7 @@ set -euo pipefail
 
 ROOT="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-PRODUCT_VERSION="5.6.0"
+PRODUCT_VERSION="5.6.1"
 LOOK_VERSION="4.47.0"
 FUTURE_CRASH_VERSION="1.2.2"